removed testing commands
